refactor: Decouple sidebar zustand stores
